refactor: regarding https://github.com/dlr-eoc/ukis-frontend-libraries/pull/43#discussion_r479376370 , moved SunlightComponent out into components directory
