+++ projects/demo-maps/src/app/components/sunlight/sunlight.component.ts
